Refactor ProvidedDependency to avoid duplication
