+++ lib/JavascriptGenerator.js
@@ -38,8 +38,8 @@ const sortByFragmentIndex = ([a, i], [b, j]) => {
 class JavascriptGenerator {
 	/**
 	 * @param {Module} module the entry module
-	 * @param {RuntimeTemplate} runtimeTemplate the runtime template
 	 * @param {DependencyTemplates} dependencyTemplates the dependency templates
+	 * @param {RuntimeTemplate} runtimeTemplate the runtime template
 	 * @returns {Source} the generated source
 	 */
 	generate(module, dependencyTemplates, runtimeTemplate) {
