+++ lib/ProvidePlugin.js
@@ -5,23 +5,10 @@
 
 "use strict";
 
-const InitFragment = require("./InitFragment");
 const { approve } = require("./JavascriptParserHelpers");
 const NullFactory = require("./NullFactory");
 const ConstDependency = require("./dependencies/ConstDependency");
-const ModuleDependency = require("./dependencies/ModuleDependency");
-
-/** @typedef {import("webpack-sources").ReplaceSource} ReplaceSource */
-/** @typedef {import("./Dependency")} Dependency */
-/** @typedef {import("./DependencyTemplates")} DependencyTemplates */
-/** @typedef {import("./RuntimeTemplate")} RuntimeTemplate */
-
-/**
- * @param {string[]} path the property path array
- * @returns {string} the converted path
- */
-const pathToString = path =>
-	path.length > 0 ? path.map(part => `[${JSON.stringify(part)}]`).join("") : "";
+const ProvidedDependency = require("./dependencies/ProvidedDependency");
 
 class ProvidePlugin {
 	constructor(definitions) {
@@ -44,7 +31,7 @@ class ProvidePlugin {
 				);
 				compilation.dependencyTemplates.set(
 					ProvidedDependency,
-					new ProvidedDependencyTemplate()
+					new ProvidedDependency.Template()
 				);
 				const handler = (parser, parserOptions) => {
 					Object.keys(definitions).forEach(name => {
@@ -84,52 +71,6 @@ class ProvidePlugin {
 					.tap("ProvidePlugin", handler);
 			}
 		);
-	}
-}
-
-class ProvidedDependency extends ModuleDependency {
-	constructor(request, identifier, path, range) {
-		super(request);
-		this.identifier = identifier;
-		this.path = path;
-		this.range = range;
-	}
-}
-
-class ProvidedDependencyTemplate extends ModuleDependency.Template {
-	/**
-	 * @param {Dependency} dependency the dependency for which the template should be applied
-	 * @param {ReplaceSource} source the current replace source which can be modified
-	 * @param {RuntimeTemplate} runtimeTemplate the runtime template
-	 * @param {DependencyTemplates} dependencyTemplates the dependency templates
-	 * @returns {void}
-	 */
-	apply(dependency, source, runtimeTemplate, dependencyTemplates) {
-		const dep = /** @type {ProvidedDependency} */ (dependency);
-		source.replace(dep.range[0], dep.range[1] - 1, dep.identifier);
-	}
-
-	/**
-	 * @param {Dependency} dependency the dependency for which the template should be applied
-	 * @param {ReplaceSource} source the current replace source which can be modified
-	 * @param {RuntimeTemplate} runtimeTemplate the runtime template
-	 * @param {DependencyTemplates} dependencyTemplates the dependency templates
-	 * @returns {InitFragment[]|null} the init fragments
-	 */
-	getInitFragments(dependency, source, runtimeTemplate, dependencyTemplates) {
-		const dep = /** @type {ProvidedDependency} */ (dependency);
-		return [
-			new InitFragment(
-				`/* provided dependency */ var ${
-					dep.identifier
-				} = ${runtimeTemplate.moduleExports({
-					module: dep.module,
-					request: dep.request
-				})}${pathToString(dep.path)};\n`,
-				1,
-				`provided ${dep.identifier}`
-			)
-		];
 	}
 }
