+++ lib/dependencies/ProvidedDependency.js
@@ -6,7 +6,6 @@
 "use strict";
 
 const InitFragment = require("../InitFragment");
-const Template = require("../Template");
 const ModuleDependency = require("./ModuleDependency");
 
 /** @typedef {import("webpack-sources").ReplaceSource} ReplaceSource */
@@ -15,28 +14,24 @@ const ModuleDependency = require("./ModuleDependency");
 /** @typedef {import("../RuntimeTemplate")} RuntimeTemplate */
 
 /**
- * @param {ProvidedDependency} dependency [TODO]
- * @returns {string} [TODO]
+ * @param {string[]|null} path the property path array
+ * @returns {string} the converted path
  */
-const getImportVarName = dependency => {
-	return `${Template.toIdentifier(dependency.name)}__WEBPACK_PROVIDED_MODULE__`;
-};
+const pathToString = path =>
+	path !== null && path.length > 0
+		? path.map(part => `[${JSON.stringify(part)}]`).join("")
+		: "";
 
 class ProvidedDependency extends ModuleDependency {
-	constructor(request, originModule, name, specifier, range) {
+	constructor(request, identifier, path, range) {
 		super(request);
-		this.originModule = originModule;
-		this.name = name;
-		this.specifier = specifier;
+		this.identifier = identifier;
+		this.path = path;
 		this.range = range;
-	}
-
-	get type() {
-		return "provided";
 	}
 }
 
-ProvidedDependency.Template = class ProvidedDependencyTemplate extends ModuleDependency.Template {
+class ProvidedDependencyTemplate extends ModuleDependency.Template {
 	/**
 	 * @param {Dependency} dependency the dependency for which the template should be applied
 	 * @param {ReplaceSource} source the current replace source which can be modified
@@ -46,17 +41,7 @@ ProvidedDependency.Template = class ProvidedDependencyTemplate extends ModuleDep
 	 */
 	apply(dependency, source, runtimeTemplate, dependencyTemplates) {
 		const dep = /** @type {ProvidedDependency} */ (dependency);
-		const binding = runtimeTemplate.exportFromImport({
-			module: dep.module,
-			originModule: dep.originModule,
-			request: dep.request,
-			importVar: getImportVarName(dep),
-			exportName: dep.specifier,
-			asiSafe: true,
-			isCall: false,
-			callContext: null
-		});
-		source.replace(dep.range[0], dep.range[1] - 1, binding);
+		source.replace(dep.range[0], dep.range[1] - 1, dep.identifier);
 	}
 
 	/**
@@ -70,18 +55,19 @@ ProvidedDependency.Template = class ProvidedDependencyTemplate extends ModuleDep
 		const dep = /** @type {ProvidedDependency} */ (dependency);
 		return [
 			new InitFragment(
-				runtimeTemplate.importStatement({
-					update: false,
+				`/* provided dependency */ var ${
+					dep.identifier
+				} = ${runtimeTemplate.moduleExports({
 					module: dep.module,
-					originModule: dep.originModule,
-					request: dep.request,
-					importVar: getImportVarName(dep)
-				}),
-				0,
-				`provided ${dep.name}`
+					request: dep.request
+				})}${pathToString(dep.path)};\n`,
+				1,
+				`provided ${dep.identifier}`
 			)
 		];
 	}
-};
+}
+
+ProvidedDependency.Template = ProvidedDependencyTemplate;
 
 module.exports = ProvidedDependency;
