+++ lib/dependencies/SystemPlugin.js
@@ -84,7 +84,6 @@ class SystemPlugin {
 								parser.state.module.context,
 								require.resolve("../../buildin/system")
 							),
-							parser.state.module,
 							"System",
 							null,
 							expr.range
