+++ lib/node/NodeSourcePlugin.js
@@ -38,8 +38,8 @@ module.exports = class NodeSourcePlugin {
 			}
 		};
 
-		const addExpression = (parser, name, module, type, suffix) => {
-			parser.hooks.expression.for(name).tap("NodeSourcePlugin", expr => {
+		const addExpression = (parser, module, identifier, specifier, type) => {
+			parser.hooks.expression.for(identifier).tap("NodeSourcePlugin", expr => {
 				if (
 					parser.state.module &&
 					parser.state.module.resource === getPathToModule(module, type)
@@ -51,9 +51,8 @@ module.exports = class NodeSourcePlugin {
 						parser.state.module.context,
 						getPathToModule(module, type)
 					),
-					parser.state.module,
-					module,
-					suffix,
+					identifier,
+					specifier ? [specifier] : null,
 					expr.range
 				);
 				dep.loc = expr.loc;
@@ -90,7 +89,6 @@ module.exports = class NodeSourcePlugin {
 										parser.state.module.context,
 										require.resolve("../../buildin/global")
 									),
-									parser.state.module,
 									"global",
 									null,
 									expr.range
@@ -101,31 +99,31 @@ module.exports = class NodeSourcePlugin {
 					}
 					if (localOptions.process) {
 						const processType = localOptions.process;
-						addExpression(parser, "process", "process", processType, null);
+						addExpression(parser, "process", "process", null, processType);
 					}
 					if (localOptions.console) {
 						const consoleType = localOptions.console;
-						addExpression(parser, "console", "console", consoleType, null);
+						addExpression(parser, "console", "console", null, consoleType);
 					}
 					const bufferType = localOptions.Buffer;
 					if (bufferType) {
-						addExpression(parser, "Buffer", "buffer", bufferType, "Buffer");
+						addExpression(parser, "buffer", "Buffer", "Buffer", bufferType);
 					}
 					if (localOptions.setImmediate) {
 						const setImmediateType = localOptions.setImmediate;
 						addExpression(
 							parser,
-							"setImmediate",
 							"timers",
-							setImmediateType,
-							"setImmediate"
+							"setImmediate",
+							"setImmediate",
+							setImmediateType
 						);
 						addExpression(
 							parser,
-							"clearImmediate",
 							"timers",
-							setImmediateType,
-							"clearImmediate"
+							"clearImmediate",
+							"clearImmediate",
+							setImmediateType
 						);
 					}
 				};
